refactor daemon bootstrap into internal package
