+++ cmd/agent-compose/daemon_compat_test.go
@@ -1,0 +1,18 @@
+package main
+
+import (
+	"net"
+	"os"
+)
+
+func isTrustedUnixSocketConn(conn net.Conn) bool {
+	unixConn, ok := conn.(*net.UnixConn)
+	if !ok {
+		return false
+	}
+	uid, err := unixSocketPeerUID(unixConn)
+	if err != nil {
+		return false
+	}
+	return uid == os.Getuid() || uid == 0
+}
