+++ internal/daemon/peercred_darwin.go
@@ -1,0 +1,29 @@
+//go:build darwin
+
+package daemon
+
+import (
+	"net"
+
+	"golang.org/x/sys/unix"
+)
+
+func unixSocketPeerUID(conn *net.UnixConn) (int, error) {
+	raw, err := conn.SyscallConn()
+	if err != nil {
+		return 0, err
+	}
+	var (
+		cred    *unix.Xucred
+		credErr error
+	)
+	if err := raw.Control(func(fd uintptr) {
+		cred, credErr = unix.GetsockoptXucred(int(fd), unix.SOL_LOCAL, unix.LOCAL_PEERCRED)
+	}); err != nil {
+		return 0, err
+	}
+	if credErr != nil {
+		return 0, credErr
+	}
+	return int(cred.Uid), nil
+}
