+++ internal/daemon/peercred_linux.go
@@ -1,0 +1,29 @@
+//go:build linux
+
+package daemon
+
+import (
+	"net"
+
+	"golang.org/x/sys/unix"
+)
+
+func unixSocketPeerUID(conn *net.UnixConn) (int, error) {
+	raw, err := conn.SyscallConn()
+	if err != nil {
+		return 0, err
+	}
+	var (
+		cred    *unix.Ucred
+		credErr error
+	)
+	if err := raw.Control(func(fd uintptr) {
+		cred, credErr = unix.GetsockoptUcred(int(fd), unix.SOL_SOCKET, unix.SO_PEERCRED)
+	}); err != nil {
+		return 0, err
+	}
+	if credErr != nil {
+		return 0, credErr
+	}
+	return int(cred.Uid), nil
+}
