+++ internal/daemon/peercred_other.go
@@ -1,0 +1,13 @@
+//go:build !linux && !darwin
+
+package daemon
+
+import (
+	"fmt"
+	"net"
+	"runtime"
+)
+
+func unixSocketPeerUID(_ *net.UnixConn) (int, error) {
+	return 0, fmt.Errorf("unix socket peer credentials not supported on %s", runtime.GOOS)
+}
